requirements.txt
crewai>=0.102.0
textblob
sqlalchemy
python-dotenv>=1.0.0
psycopg2-binary

# Web Scraping and Parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
newspaper3k>=0.2.8
lxml[html_clean]>=5.3.0
selenium>=4.18.1

# Social Media
twikit>=0.1.0
aiohttp>=3.9.0
Mastodon.py>=1.8.1

# News and Content
newsapi-python>=0.2.7

# AI and Language Processing
openai>=1.6.1
nltk>=3.8.1

# Authentication
python-jose>=3.3.0

# Database
pydantic>=2.5.0

# Text Processing
textblob>=0.17.1

# Additional dependencies
numpy>=1.21.0
scipy>=1.7.0

# FastAPI dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
email-validator>=1.1.3

# Added this line
litellm==1.60.2